% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmat_AR1.R
\name{corrmat_AR1}
\alias{corrmat_AR1}
\title{AR(1) correlation matrix}
\usage{
corrmat_AR1(m, rho)
}
\arguments{
\item{m}{An integer. Dimension of the correlation matrix.}

\item{rho}{A number. Correlation coefficient between \eqn{(0,1)}}
}
\value{
A correlation matrix
}
\description{
An easy function to generate a AR(1) correlation matrix.
}
\examples{
corrmat_AR1(
  m = 3,
  rho = 0.2
)
}
