% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrmat_CS.R
\name{corrmat_CS}
\alias{corrmat_CS}
\title{Compound symmetric correlation matrix}
\usage{
corrmat_CS(m, rho)
}
\arguments{
\item{m}{An integer. Dimension of the correlation matrix.}

\item{rho}{A number. Correlation coefficient between \eqn{(0,1)}}
}
\value{
A correlation matrix
}
\description{
An easy function to generate a compound symmetric correlation matrix
}
\examples{
corrmat_CS(
  m = 3,
  rho = 0.2
)
}
