% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrMCT-package.R
\docType{package}
\name{corrMCT-package}
\alias{corrMCT}
\alias{corrMCT-package}
\title{corrMCT: Correlated Weighted Hochberg}
\description{
Perform additional multiple testing procedure methods to p.adjust(), such as weighted Hochberg (Tamhane, A. C., & Liu, L., 2008) \doi{10.1093/biomet/asn018}, ICC adjusted Bonferroni method (Shi, Q., Pavey, E. S., & Carter, R. E., 2012) \doi{10.1002/pst.1514} and a new correlation corrected weighted Hochberg for correlated endpoints.
}
\author{
\strong{Maintainer}: Xin-Wei Huang \email{xinweihuangstat@gmail.com} (\href{https://orcid.org/0000-0003-4238-3081}{ORCID})

Other contributors:
\itemize{
  \item Jia Hua [contributor]
  \item Bhramori Banerjee [contributor]
  \item Xuelong Wang [contributor]
  \item Qing Li [contributor]
  \item Merck & Co., Inc [copyright holder, funder]
}

}
\keyword{internal}
